+++ app/PackageBuilder/PackageBuilder.Domain.Entities/DataProviders/Events/DataProviderCreated.cs
@@ -10,13 +10,13 @@ namespace PackageBuilder.Domain.Entities.DataProviders.Events
     {
         public readonly DataProviderName Name;
 		public readonly string Description;
-        public readonly double CostPrice;
+        public readonly decimal CostPrice;
         public readonly Type ResponseType;
         public readonly string Owner;
         public readonly DateTime CreatedDate;
         public readonly IEnumerable<IDataField> RequestFields;
         public readonly IEnumerable<IDataField> DataFields;
-        public DataProviderCreated(Guid id, DataProviderName name, string description, double costPrice, Type responseType, string owner, DateTime createdDate, IEnumerable<IDataField> requestFields, IEnumerable<IDataField> dataFields)
+        public DataProviderCreated(Guid id, DataProviderName name, string description, decimal costPrice, Type responseType, string owner, DateTime createdDate, IEnumerable<IDataField> requestFields, IEnumerable<IDataField> dataFields)
         {
 			Id = id;
 			Name = name;
