+++ app/PackageBuilder/PackageBuilder.Domain.Entities/Packages/Commands/UpdatePackage.cs
@@ -11,7 +11,7 @@ namespace PackageBuilder.Domain.Entities.Packages.Commands
     {
         public readonly int Version;
 
-        public UpdatePackage(Guid id, string name, string description, double costPrice, double salePrice, string notes, IEnumerable<Industry> industries, State state, int version, string owner, DateTime createdDate, DateTime editedDate, IEnumerable<IDataProviderOverride> dataProviderValueOverrides)
+        public UpdatePackage(Guid id, string name, string description, decimal costPrice, decimal salePrice, string notes, IEnumerable<Industry> industries, State state, int version, string owner, DateTime createdDate, DateTime editedDate, IEnumerable<IDataProviderOverride> dataProviderValueOverrides)
             : base(id, name, description, costPrice, salePrice, notes, industries, state, owner, createdDate, editedDate, dataProviderValueOverrides)
         {
             Version = version;
